Switch to bigint

+++ dbt/models/reporting/reporting.vw_ratio_curve.sql
@@ -49,9 +49,9 @@ SELECT
     vpu.pin,
     vpu.township_name,
     vpu.neighborhood_number,
-    CAST(vpv.pre_mailed_tot * 10 AS INT) AS desk_review_value,
-    CAST(model_vals.model_value AS INT) AS model_value,
-    CAST(model_vals.sale_price AS INT) AS sale_price,
+    CAST(vpv.pre_mailed_tot * 10 AS BIGINT) AS desk_review_value,
+    CAST(model_vals.model_value AS BIGINT) AS model_value,
+    CAST(model_vals.sale_price AS BIGINT) AS sale_price,
     model_vals.sale_date,
     model_vals.sale_document_number
 FROM most_recent_pin AS vpu
